move activity-rectangle content into own folder
closes #324
