+++ projects/ngx-d3timeline/src/lib/activity-rectangle/activity-drag.service.ts
