+++ projects/ngx-d3timeline/src/lib/activity-rectangle/activity-rectangle.component.spec.ts
