+++ projects/ngx-d3timeline/src/lib/activity-rectangle/selectors/activity-rectangle-position.selectors.ts
@@ -12,7 +12,7 @@ import {
   getActivityTopLeft,
   getActivityTransform,
   getOffsetActivityTopLeft
-} from '../content-utils';
+} from '../../content/content-utils';
 import { MemoizedSelector } from '../../store-lib/selector/memoized-selector';
 import { Point } from '../../core/point';
