+++ projects/ngx-d3timeline/src/lib/activity-rectangle/selectors/activity-rectangle-size.selectors.ts
@@ -8,7 +8,7 @@ import {
   getRectBreadthInTimeAxis,
   getRectHeight,
   getRectWidth
-} from '../content-utils';
+} from '../../content/content-utils';
 
 export const selectRectBreadthInTimeAxis = createSelector(
   selectTimeScale,
