+++ projects/ngx-d3timeline/src/lib/content/content-utils.ts
@@ -2,7 +2,7 @@ import { PositionedActivity } from '../activity/positioned-activity';
 import { TimeScale, BandScale } from '../scale-types';
 import { Orientation } from '../core/orientation';
 import { Point, translatePoint, pointToTransform } from '../core/point';
-import { ActivityRectangle } from './activity-rectangle';
+import { ActivityRectangle } from '../activity-rectangle/activity-rectangle';
 import { getTextWidth } from '../core/text-utils';
 
 type PositionInAxis = (p: PositionedActivity) => number;
